+++ atlasdb-client/src/main/java/com/palantir/atlasdb/table/description/ColumnMetadataDescription.java
@@ -16,7 +16,9 @@
 package com.palantir.atlasdb.table.description;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
+import java.util.SortedMap;
 
 import javax.annotation.concurrent.Immutable;