+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/cleaner/KeyValueServicePuncherStore.java
@@ -26,17 +26,11 @@ import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.api.RangeRequest;
 import com.palantir.atlasdb.keyvalue.api.RowResult;
 import com.palantir.atlasdb.keyvalue.api.Value;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence;
 import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.ValueByteOrder;
 import com.palantir.atlasdb.ptobject.EncodingUtils;
-import com.palantir.atlasdb.table.description.ColumnMetadataDescription;
-import com.palantir.atlasdb.table.description.ColumnValueDescription;
-import com.palantir.atlasdb.table.description.NameComponentDescription;
 import com.palantir.atlasdb.table.description.NameMetadataDescription;
-import com.palantir.atlasdb.table.description.NamedColumnDescription;
 import com.palantir.atlasdb.table.description.TableMetadata;
 import com.palantir.atlasdb.table.description.ValueType;
-import com.palantir.atlasdb.transaction.api.ConflictHandler;
 import com.palantir.common.base.ClosableIterator;
 
 /**
@@ -83,17 +77,11 @@ public final class KeyValueServicePuncherStore implements PuncherStore {
     }
 
     private void tryInitialize() {
-        keyValueService.createTable(AtlasDbConstants.PUNCH_TABLE, new TableMetadata(
-                NameMetadataDescription.create(ImmutableList.of(
-                        new NameComponentDescription.Builder()
-                                .componentName("time")
-                                .type(ValueType.VAR_LONG)
-                                .byteOrder(ValueByteOrder.DESCENDING)
-                                .build())),
-                new ColumnMetadataDescription(ImmutableList.of(
-                        new NamedColumnDescription("t", "t", ColumnValueDescription.forType(ValueType.VAR_LONG)))),
-                ConflictHandler.IGNORE_ALL,
-                TableMetadataPersistence.LogSafety.SAFE).persistToBytes());
+        keyValueService.createTable(AtlasDbConstants.PUNCH_TABLE, TableMetadata.internal()
+                .rowMetadata(NameMetadataDescription.create("time", ValueType.VAR_LONG, ValueByteOrder.DESCENDING))
+                .singleNamedColumn("t", "t", ValueType.VAR_LONG)
+                .build()
+                .persistToBytes());
     }
 
     @Override